【バック】コントローラーのディレクトリ構成の変更
